+++ ports/extremetuxracer/src/etr_types.h
@@ -14,6 +14,8 @@ but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 ---------------------------------------------------------------------*/
+// PSP port modifications, 2026-09-07. See docs/porting.md in the port repository.
+
 
 #ifndef ETR_TYPES_H
 #define ETR_TYPES_H
@@ -50,7 +52,7 @@ struct TGameData {
 	TToolMode toolmode;
 	float time_step;
 	TGameType game_type;
-	double finish_brake;
+	float finish_brake;
 	int argument;
 	int treesize;
 	int treevar;
