+++ ports/extremetuxracer/src/matrices.h
@@ -13,6 +13,8 @@ but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 ---------------------------------------------------------------------*/
+// PSP port modifications, 2026-09-07. See docs/porting.md in the port repository.
+
 
 #ifndef MATRICES_H
 #define MATRICES_H
@@ -21,19 +23,19 @@ GNU General Public License for more details.
 
 template<int ix, int iy>
 class TMatrix {
-	double _data[ix][iy];
+	float _data[ix][iy];
 public:
 	constexpr TMatrix() = default;
 	TMatrix(const TVector3d& w1, const TVector3d& w2, const TVector3d& w3);
 
-	double* operator[](int index) { return _data[index]; }
-	constexpr const double* operator[](int index) const { return _data[index]; }
-	constexpr const double* data() const { return (double*)_data; }
+	float* operator[](int index) { return _data[index]; }
+	constexpr const float* operator[](int index) const { return _data[index]; }
+	constexpr const float* data() const { return (float*)_data; }
 
 	void SetIdentity();
-	void SetRotationMatrix(double angle, char axis);
-	void SetTranslationMatrix(double x, double y, double z);
-	void SetScalingMatrix(double x, double y, double z);
+	void SetRotationMatrix(float angle, char axis);
+	void SetTranslationMatrix(float x, float y, float z);
+	void SetScalingMatrix(float x, float y, float z);
 
 	TMatrix GetTransposed() const;
